[RSPEED-1400] Upcoming potentially affected systems
* add potenatially affected systems data provided by API
* use the correct type definition
* cleanup of unused mock data
