+++ src/Components/UpcomingRow/UpcomingRow.tsx
@@ -1,10 +1,10 @@
 import '@patternfly/react-core/dist/styles/base.css';
-import { Record, columnNames } from '../Upcoming/mock_data';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon';
 import ExclamationTriangleIcon from '@patternfly/react-icons/dist/esm/icons/exclamation-triangle-icon';
 import InfoCircleIcon from '@patternfly/react-icons/dist/esm/icons/info-circle-icon';
 import React, { lazy } from 'react';
 import { Tbody, Td, Tr } from '@patternfly/react-table';
+import { UpcomingChanges } from '../../types/UpcomingChanges';
 import {
   Button,
   Icon,
@@ -17,10 +17,16 @@ import {
   TextVariants,
 } from '@patternfly/react-core';
 const LifecycleModalWindow = lazy(() => import('../../Components/LifecycleModalWindow/LifecycleModalWindow'));
-import { SYSTEM_ID } from '../../__mocks__/mockData';
+
+export const columnNames = {
+  name: 'Name',
+  type: 'Type',
+  release: 'Release',
+  date: 'Date',
+};
 
 interface TableRowProps {
-  repo: Record;
+  repo: UpcomingChanges;
   columnNames: typeof columnNames;
   rowIndex: number;
   isExpanded: boolean;
@@ -143,7 +149,7 @@ export const TableRow: React.FunctionComponent<TableRowProps> = ({
                         onClick={(event) => {
                           handleModalToggle(event);
                           setModalDataName(String(repo.name));
-                          setModalData(SYSTEM_ID);
+                          setModalData(repo.details?.potentiallyAffectedSystems);
                         }}
                         style={{
                           marginTop: '-4px',
@@ -151,7 +157,7 @@ export const TableRow: React.FunctionComponent<TableRowProps> = ({
                           marginLeft: '-16px',
                         }}
                       >
-                        {repo.details.potentiallyAffectedSystems}
+                        {repo.details.potentiallyAffectedSystemsCount}
                       </Button>
                     </TextListItem>
                     <TextListItem component={TextListItemVariants.dt}>Tracking ticket</TextListItem>
